Clang-tidy CI test: bump version from 15 to 16 (#5592)
This PR bumps the version used for `clang-tidy` CI tests from 15 to 16.
It also addresses all the issues found with the upgraded tool.

Most of the issues are related to this newly introduced check:
-
[cppcoreguidelines-avoid-const-or-ref-data-members](https://releases.llvm.org/16.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/cppcoreguidelines/avoid-const-or-ref-data-members.html)
The check enforces [CppCoreGuidelines about constant and reference data
members](
https://github.com/isocpp/CppCoreGuidelines/blob/master/CppCoreGuidelines.md#c12-dont-make-data-members-const-or-references)
.
In general, I understand the argument against using constant or
reference data members.
~~There is however one case in which I am not fully convinced by the
suggestion of the tool: in
[PML.H](https://github.com/ECP-WarpX/WarpX/pull/5592/files#diff-f1e020ebe3cd2222f399d50ff05769d0c70482f0e12bbe29b498e9ab2d0f4a53)
`const amrex::BoxArray& m_grids;` becomes `amrex::BoxArray m_grids;` and
I am wondering if this can be an issue for performances. Maybe we could
consider using a (possibly smart) pointer to the `BoxArray`, instead of
making a copy.~~ (we are now using a pointer for `amrex::BoxArray
m_grids;`).

Few issues were instead related to these checks:
-
[modernize-loop-convert](https://releases.llvm.org/16.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/modernize/loop-convert.html)
This check was already enabled, but `clang-tidy-16` broadens its scope
with respect to previous versions.

-
[modernize-use-auto](https://releases.llvm.org/16.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/modernize/use-auto.html)
Only one case. I am a bit confused because this should have been found
also by version 15 of the tool.

-
[misc-use-anonymous-namespace](https://releases.llvm.org/16.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/misc/use-anonymous-namespace.html)
This is a new check. Actually, the issues found with this check were
false positives, but they disappeared when I properly set
`misc-use-anonymous-namespace.HeaderFileExtensions` in the `.clang-tidy`
configuration file to recognize `.H` files as headers.

-
[misc-misplaced-const](https://releases.llvm.org/16.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/misc/misplaced-const.html)
Only one case. I am a bit confused because this should have been found
also by version 15 of the tool.

-
[readability-misleading-indentation](https://releases.llvm.org/16.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/readability/misleading-indentation.html)
[**NOW DISABLED DUE TO FALSE POSITIVES**]
This check was already enabled. However, with this newer version of the
tool, it seems to me that it generates some false positives. Therefore,
I would like to propose to **disable** it. We may try to re-enable it
when we will bump the version from 16 to 17.