Added CI to test secondary ion emission in RZ. (#5576)
This PR adds secondary ion emission through a callback function,
allowing secondary electrons to be emitted when an ion hits the embedded
boundary. In the following CI test, the random seed was fixed to ensure
consistent emission of secondary electrons for reproducibility. We used
a secondary electron emission yield (SEY) of 0.4.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Remi Lehe <remi.lehe@normalesup.org>