Add FieldPoyntingFlux reduced diagnostic (#5475)
This adds a reduced diagnostic that calculates the Poynting flux on the
surfaces of the domain, providing the power flow into and out of the
domain. This also includes the time integrated data.

When using the implicit evolve scheme, to get the energy accounting
correct, the flux needs to be calculated at the mid step. For this
reason, the `ComputeDiagsMidStep` was added which is called directly at
the appropriate times.

Because of the time integration, there are two main differences of this
reduced diagnostic compared to the others. The first is that it is
calculated every time step in order to get the full resolution in time.
The intervals parameter still controls how often the diagnostic data is
written out. The second is that a facility is added to write out the
values of the time integration to a file when a checkpoint is made, so
on a restart the integration can continue with the previous values. The
facility was written in a general way so that other reduced diagnostics
can also do this.

The CI test using the implicit solver is dependent on PR #5498 and PR
#5489.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>