Add PECInsulator to Curl-Curl BC (#5667)
This is a temporary fix, setting a boundary condition for the Curl-Curl
preconditioner for the implicit solver. This now sets the BC to
Dirichlet for the PEC regions. A correct solution would have to be
implemented in the curl-curl solver because of the split between the PEC
and insulator sections.