WarpX class: fuse together doFieldIonization with doFieldIonization(lev) and doQEDEvents with doQEDEvents(lev) (#5671)
`doFieldIonization(lev) ` is called only once, inside
`doFieldIonization` , which is simply a loop over the levels calling for
each level `doFieldIonization(lev) `. The same happens for
`doQEDEvents`.

In order to simplify the interface of the WarpX class, I would like to
propose to drop `doFieldIonization(lev) ` and `doQEDEvents(lev) `, and
to integrate their code respectively in `doFieldIonization` and
`doQEDEvents`.