Small fix in Perlmutter GPU sbatch script  (#5683)
Changes in Perlmutter GPU job script:
from `#SBATCH --cpus-per-task=16` to `#SBATCH --cpus-per-task=32`.

This is to request (v)cores in consecutive blocks.

GPU 3 is closest to CPU cores 0-15, 64-79, 
GPU 2 to CPU cores 16-31, 80-95, 
...

If `--cpus-per-task=16`, MPI ranks 0 and 1 are mapped to cores 0 and 8.
If `--cpus-per-task=32`, MPI ranks 0 and 1 are mapped to cores 0 and 16.

Visual representation

![pm_gpu_vcores_mpi](https://github.com/user-attachments/assets/edf0721f-7321-49ab-bf37-4b55a7c422cc)

---------

Co-authored-by: Axel Huebl <axel.huebl@plasma.ninja>